Remove physics.World class, move logic into helpers for a pymunk.Space object
There's no need for a wrapper around pymunk.Space yet, and if there is it should
probably go in the skeleton layer.  Composable helpers are much nicer to work with
than a monolithic wrapper that forces conventions on the user.  The rapid layer
should avoid locking the user in as much as possible.